Set utility functions in better importable modules
