+++ datalad_gooey/utils.py
@@ -1,9 +1,20 @@
 from pathlib import Path
+
 from PySide6.QtUiTools import QUiLoader
 from PySide6.QtCore import (
     QFile,
     QIODevice,
 )
+
+
+class _NoValue:
+    """Type to annotate the absence of a value
+
+    For example in a list of parameter defaults. In general `None` cannot
+    be used, as it may be an actual value, hence we use a local, private
+    type.
+    """
+    pass
 
 
 def load_ui(name, parent=None):
